some more variable renames and made the fog of war fade radially from the camera rather than distance along the z-axis

+++ src/client/CameraModel.java
@@ -15,7 +15,7 @@ public class CameraModel {
     protected int zMax;
     protected Model model;
     protected int surface;
-    protected int drawOrderVal;
+    protected double drawOrderVal;
     protected int xNormal;
     protected int zNormal;
     protected int yNormal;
